work towards a sensible public interface
