+++ src/c/retransmission/handlers.c
@@ -246,3 +246,275 @@ int handle_retrdata(struct rasta_connection *connection, struct RastaPacket *rec
 
     return result;
 }
+
+// HACK
+// TODO: Also fill this from kex handlers
+struct rasta_connection *handle_conreq(struct rasta_connection *connection, struct RastaPacket *receivedPacket) {
+    logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Received ConnectionRequest from %d", receivedPacket->sender_id);
+
+    if (connection->current_state == RASTA_CONNECTION_CLOSED || connection->current_state == RASTA_CONNECTION_DOWN) {
+        sr_init_connection(connection, RASTA_ROLE_SERVER);
+
+        // initialize seq num
+        connection->sn_t = connection->sn_i = receivedPacket->sequence_number;
+
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Using %lu as initial sequence number",
+                   (long unsigned int)connection->sn_t);
+
+        connection->current_state = RASTA_CONNECTION_DOWN;
+
+        // check received packet (5.5.2)
+        if (!sr_check_packet(connection, connection->logger, &connection->config->sending, receivedPacket, "RaSTA HANDLE: ConnectionRequest")) {
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Packet is not valid");
+            sr_close_connection(connection, RASTA_DISC_REASON_PROTOCOLERROR, 0);
+            return connection;
+        }
+
+        // received packet is a ConReq -> check version
+        struct RastaConnectionData connectionData = extractRastaConnectionData(receivedPacket);
+
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Client has version %.4s", connectionData.version);
+
+        if (compare_version(&RASTA_VERSION, &connectionData.version) == 0 ||
+            compare_version(&RASTA_VERSION, &connectionData.version) == -1 ||
+            version_accepted(connection->config, &connectionData.version)) {
+
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Version accepted");
+
+            // same version, or lower version -> client has to decide -> send ConResp
+
+            // set values according to 5.6.2 [3]
+            update_connection_attrs(connection, receivedPacket);
+            update_confirmed_attrs(connection, receivedPacket);
+
+            // save N_SENDMAX of partner
+            connection->connected_recv_buffer_size = connectionData.send_max;
+
+            connection->t_i = connection->config->sending.t_max;
+
+            unsigned char *version = (unsigned char *)RASTA_VERSION;
+
+            // send ConResp
+            struct RastaPacket conresp = createConnectionResponse(connection->remote_id, connection->my_id,
+                                                                  connection->sn_t, connection->cs_t,
+                                                                  cur_timestamp(), connection->cts_r,
+                                                                  connection->config->sending.send_max,
+                                                                  version, &connection->redundancy_channel->hashing_context);
+
+            connection->sn_t = connection->sn_t + 1;
+
+            connection->current_state = RASTA_CONNECTION_START;
+
+            fire_on_connection_state_change(sr_create_notification_result(NULL, connection));
+
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Send Connection Response - waiting for Heartbeat");
+
+            // Send connection response immediately (don't go through packet batching)
+            redundancy_mux_send(connection->redundancy_channel, &conresp, connection->role);
+
+            freeRastaByteArray(&conresp.data);
+        } else {
+            logger_log(connection->logger, LOG_LEVEL_INFO, "RaSTA HANDLE: ConnectionRequest", "Version unacceptable - sending DisconnectionRequest");
+            sr_close_connection(connection, RASTA_DISC_REASON_INCOMPATIBLEVERSION, 0);
+            return connection;
+        }
+    } else {
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionRequest", "Connection is in invalid state (%d) send DisconnectionRequest", connection->current_state);
+        sr_close_connection(connection, RASTA_DISC_REASON_UNEXPECTEDTYPE, 0);
+    }
+    return connection;
+}
+
+struct rasta_connection *handle_conresp(struct rasta_connection *con, struct RastaPacket *receivedPacket) {
+
+    logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Received ConnectionResponse from %d", receivedPacket->sender_id);
+
+    logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Checking packet..");
+    if (!sr_check_packet(con, con->logger, &con->config->sending, receivedPacket, "RaSTA HANDLE: ConnectionResponse")) {
+        sr_close_connection(con, RASTA_DISC_REASON_PROTOCOLERROR, 0);
+        return con;
+    }
+
+    if (con->current_state == RASTA_CONNECTION_START) {
+        if (con->role == RASTA_ROLE_CLIENT) {
+            // handle normal conresp
+            logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Current state is in order");
+
+            // correct type of packet received -> version check
+            struct RastaConnectionData connectionData = extractRastaConnectionData(receivedPacket);
+
+            // logger_log(&connection->logger, LOG_LEVEL_INFO, "RaSTA open con", "server is running RaSTA version %s", connectionData.version);
+
+            logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Client has version %s", connectionData.version);
+
+            if (version_accepted(con->config, &connectionData.version)) {
+
+                logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Version accepted");
+
+                // same version or accepted versions -> send hb to complete handshake
+
+                // set values according to 5.6.2 [3]
+                update_connection_attrs(con, receivedPacket);
+                con->cs_r = receivedPacket->confirmed_sequence_number;
+
+                // update state, ready to send data
+                con->current_state = RASTA_CONNECTION_UP;
+
+                // send hb
+                logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Sending heartbeat..");
+                sendHeartbeat(con, 1);
+
+#ifdef ENABLE_OPAQUE
+                if (con->config->kex.mode == KEY_EXCHANGE_MODE_OPAQUE) {
+                    send_KexRequest(con);
+                }
+#endif
+
+                // fire connection state changed event
+                fire_on_connection_state_change(sr_create_notification_result(NULL, con));
+                // fire handshake complete event
+                fire_on_handshake_complete(sr_create_notification_result(NULL, con));
+
+                // start sending heartbeats
+                enable_timed_event(&con->send_heartbeat_event);
+
+                con->hb_locked = 0;
+
+                // save the N_SENDMAX of remote
+                con->connected_recv_buffer_size = connectionData.send_max;
+
+                // arm the timeout timer
+                enable_timed_event(&con->timeout_event);
+
+            } else {
+                // version not accepted -> disconnect
+                logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Version not acceptable - send DisonnectionRequest");
+                sr_close_connection(con, RASTA_DISC_REASON_INCOMPATIBLEVERSION, 0);
+                return con;
+            }
+        } else {
+            // Server don't receive conresp
+            sr_close_connection(con, RASTA_DISC_REASON_UNEXPECTEDTYPE, 0);
+
+            logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Server received ConnectionResponse - send Disconnection Request");
+            return con;
+        }
+    } else if (con->current_state == RASTA_CONNECTION_RETRREQ || con->current_state == RASTA_CONNECTION_RETRRUN || con->current_state == RASTA_CONNECTION_UP) {
+        sr_close_connection(con, RASTA_DISC_REASON_UNEXPECTEDTYPE, 0);
+        logger_log(con->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: ConnectionResponse", "Received ConnectionResponse in wrong state - semd DisconnectionRequest");
+        return con;
+    }
+    return con;
+}
+
+int handle_hb(rasta_connection *connection, struct RastaPacket *receivedPacket) {
+    logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "Received heartbeat from %d", receivedPacket->sender_id);
+
+    if (connection->current_state == RASTA_CONNECTION_START) {
+        // heartbeat is for connection setup
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "Establish connection");
+
+        // if SN not in Seq -> disconnect and close connection
+        if (!sr_sn_in_seq(connection, receivedPacket)) {
+            logger_log(connection->logger, LOG_LEVEL_INFO, "RaSTA HANDLE: Heartbeat", "Connection HB SN not in Seq");
+
+            if (connection->role == RASTA_ROLE_SERVER) {
+                // SN not in Seq
+                sr_close_connection(connection, RASTA_DISC_REASON_SEQNERROR, 0);
+            } else {
+                // Client
+                sr_close_connection(connection, RASTA_DISC_REASON_UNEXPECTEDTYPE, 0);
+            }
+        }
+
+        // if client receives HB in START -> disconnect and close
+        if (connection->role == RASTA_ROLE_CLIENT) {
+            sr_close_connection(connection, RASTA_DISC_REASON_UNEXPECTEDTYPE, 0);
+        }
+
+        if (sr_cts_in_seq(connection, &connection->config->sending, receivedPacket)) {
+            // set values according to 5.6.2 [3]
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "Heartbeat is valid connection successful");
+            update_connection_attrs(connection, receivedPacket);
+            connection->cs_r = receivedPacket->confirmed_sequence_number;
+
+            if (connection->config->kex.mode == KEY_EXCHANGE_MODE_NONE) {
+                // sequence number correct, ready to receive data
+                connection->current_state = RASTA_CONNECTION_UP;
+            } else {
+                // need to negotiate session key first
+                connection->current_state = RASTA_CONNECTION_KEX_REQ;
+            }
+
+            connection->hb_locked = 0;
+
+            // fire connection state changed event
+            fire_on_connection_state_change(sr_create_notification_result(NULL, connection));
+            // fire handshake complete event
+            fire_on_handshake_complete(sr_create_notification_result(NULL, connection));
+
+            // start sending heartbeats
+            enable_timed_event(&connection->send_heartbeat_event);
+
+            // arm the timeout timer
+            enable_timed_event(&connection->timeout_event);
+
+            connection->is_new = true;
+
+            return 1;
+        } else {
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "Heartbeat is invalid");
+
+            // sequence number check failed -> disconnect
+            sr_close_connection(connection, RASTA_DISC_REASON_PROTOCOLERROR, 0);
+            return 0;
+        }
+    }
+
+    if (sr_sn_in_seq(connection, receivedPacket)) {
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "SN in SEQ");
+        // heartbeats also permissible during key exchange phase, since computation could exceed heartbeat interval
+        if (connection->current_state == RASTA_CONNECTION_UP || connection->current_state == RASTA_CONNECTION_RETRRUN ||
+            connection->current_state == RASTA_CONNECTION_KEX_REQ || connection->current_state == RASTA_CONNECTION_KEX_RESP ||
+            connection->current_state == RASTA_CONNECTION_KEX_AUTH) {
+            // check cts_in_seq
+            if (sr_cts_in_seq(connection, &connection->config->sending, receivedPacket)) {
+                logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "CTS in SEQ");
+
+                updateTimeoutInterval(receivedPacket->confirmed_timestamp, connection, &connection->config->sending);
+                updateDiagnostic(connection, receivedPacket, &connection->config->sending);
+
+                // set values according to 5.6.2 [3]
+                update_connection_attrs(connection, receivedPacket);
+                update_confirmed_attrs(connection, receivedPacket);
+
+                // cs_r updated, remove confirmed messages
+                sr_remove_confirmed_messages(connection);
+
+                if (connection->current_state == RASTA_CONNECTION_RETRRUN) {
+                    connection->current_state = RASTA_CONNECTION_UP;
+                    logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "State changed from RetrRun to Up");
+                    fire_on_connection_state_change(sr_create_notification_result(NULL, connection));
+                }
+            } else {
+                logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "CTS not in SEQ - send DisconnectionRequest");
+                // close connection
+                sr_close_connection(connection, RASTA_DISC_REASON_TIMEOUT, 0);
+            }
+        }
+    } else {
+        logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "SN not in SEQ");
+
+        if (connection->current_state == RASTA_CONNECTION_UP || connection->current_state == RASTA_CONNECTION_RETRRUN) {
+            // ignore message, send RetrReq and goto state RetrReq
+            // TODO:send retransmission
+            // send_retrreq(con);
+            connection->current_state = RASTA_CONNECTION_RETRREQ;
+            logger_log(connection->logger, LOG_LEVEL_DEBUG, "RaSTA HANDLE: Heartbeat", "Send retransmission");
+
+            // fire connection state changed event
+            fire_on_connection_state_change(sr_create_notification_result(NULL, connection));
+        }
+    }
+    return 0;
+}
